Move validator client exceptions into a exception package
