+++ validator/client/src/test/java/tech/pegasys/teku/validator/client/RetryingDutyLoaderTest.java
@@ -26,6 +26,7 @@ import tech.pegasys.teku.infrastructure.async.SafeFuture;
 import tech.pegasys.teku.infrastructure.async.StubAsyncRunner;
 import tech.pegasys.teku.validator.api.NodeSyncingException;
 import tech.pegasys.teku.validator.client.duties.ScheduledDuties;
+import tech.pegasys.teku.validator.client.exception.NodeDataUnavailableException;
 
 class RetryingDutyLoaderTest {
